requirements.txt
asgiref==3.2.5
astroid==2.3.3
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-phonenumber-field==4.0.0
django-qr-code==1.1.0
httplib2==0.17.0
idna==2.9
isort==4.3.21
jedi==0.16.0
jsonpickle==1.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
parso==0.6.2
paytm-pg==1.0.6
pexpect==4.8.0
phonenumberslite==8.11.5
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prompt-toolkit==3.0.4
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.12.0
python3-openid==3.1.0
pytz==2019.3
pyzmq==19.0.0
qrcode==6.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
xmltodict==0.12.0
